\name{CUM.LN}
\alias{CUM.LN}
\docType{data}
\title{Cumulative distribution of the Ln statistic}
\description{
Cumulative distribution of the statistic "size of the longest increasing subsequence", it is used in the LIS.pvalue function.
}
\usage{data(CUM.LN)}
\format{
  The format is:
 num [1:90, 1:90] 1.00000 0.50000 0.16667 0.04167 0.00833 ...
}
\details{
}
\source{
}
\references{
}
\examples{
data(CUM.LN)
## maybe str(CUM.LN) ; plot(CUM.LN) ...
}
\keyword{internal}
