\name{LIStest-package}
\alias{LIStest-package}
\alias{LIStest}
\docType{package}
\title{Tests of independence based on the Longest Increasing Subsequence}
\description{
Tests for independence between X and Y computed from a paired sample (x1,y1), ..., (xn,yn) of (X,Y), using one of the following statistics (a) the Longest Increasing Subsequence (Ln), (b) JLn, a Jackknife version of Ln or (c) JLMn, a Jackknife version of the longest monotonic subsequence. This family of tests can be applied under the assumption of continuity of X and Y.
}
\details{
\tabular{ll}{
Package: \tab LIStest\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2014-03-12\cr
License: \tab GPL-2\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
J. E. Garcia and V. A. Gonzalez-Lopez
Maintainer: J. E. Garcia  <jg@ime.unicamp.br>
%~~ The author and/or maintainer of the package ~~
}
\references{ J. E. Garcia, V. A. Gonzalez-Lopez, Independence tests for 
continuous random variables based on the longest increasing subsequence, 
Journal of Multivariate Analysis (2014), \url{http://dx.doi.org/10.1016/j.jmva.2014.02.010 }}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
%~~ in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~}
%\examples{
%~~ simple examples of the most important functions ~~}
