\name{MODE.LN}
\alias{MODE.LN}
\docType{data}
\title{ Mode of the statistic Ln  (Integer)}
\description{
 Mode of the statistic "size of the longest increasing subsequence", it is used in the LIS.pvalue function to find the p-value on the two sided alternative. 
}
\usage{data(MODE.LN)}
\format{
  The format is:
 int [1:69] 1 1 2 2 3 3 3 4 4 4 ...
}
\details{
}
\source{ 
A nonparametric idependence test for small sample size
}
\references{
}
\examples{
data(MODE.LN)
## maybe str(MODE.LN) ; plot(MODE.LN) ...
}
\keyword{internal}
